refactor: split snapshot attempt cleanup module
